Refactor Complete Task Function
