+++ src/components/task-progress/TaskCard.tsx
@@ -3,6 +3,7 @@ import type { Task } from '@/types'
 import TaskIcon from '../TaskIcon'
 import { useRecoilState } from 'recoil'
 import { tasksState } from '@/features/taskAtoms'
+import { useTasksAction } from '@/hooks/useTasksAction'
 
 
 interface TaskCardProps {
@@ -10,20 +11,9 @@ interface TaskCardProps {
 }
 
 const TaskCard = ({ task }: TaskCardProps): JSX.Element => {
+  const { moveTaskCard } = useTasksAction()
+  
   const isStarted = task.progressOrder === TASK_PROGRESS_ID.NOT_STARTED
-  const [tasks, setTasks] = useRecoilState<Task[]>(tasksState)
-
-  const moveTaskCard = (taskId: number, directionNumber: 1 | -1): void => {
-    
-     
-    const updatedTasks: Task[] = tasks.map((item) =>
-      item.id === taskId ? { ...item, progressOrder: item.progressOrder+directionNumber } : item,
-    )
-    setTasks(updatedTasks)
-    console.log(tasks)
-    console.log(updatedTasks)
-
-  }
 
   return (
     <div className="bg-green-200 p-6 rounded-xl my-2 flex flex-col gap-y-2 text-xl relative">
